roleDTO added and security configuration updated with auth2

+++ src/main/java/com/server/ecommerceapp/repository/IProductRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface IProductRepository extends JpaRepository<Product, Integer> {
+public interface IProductRepository extends JpaRepository<Product, Long> {
     Optional<Product> findProductByTitle(String Title);
 
 }
